[WIP] - break up entry point and file monitoring into seperate classes
