+++ src/integration-test/java/com/loucans/bob/csvtojson/CsvToJsonIntegrationTests.java
@@ -4,19 +4,12 @@ import com.loucans.bob.csvtojson.exception.InvalidDataRowException;
 import com.loucans.bob.csvtojson.processor.CsvProcessor;
 import com.loucans.bob.csvtojson.processor.ErrorLogger;
 import com.loucans.bob.csvtojson.processor.OutputWriter;
-import org.apache.commons.lang3.StringUtils;
 import org.json.JSONArray;
-import org.json.JSONObject;
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.Mockito;
-import org.mockito.runners.MockitoJUnitRunner;
 
 import java.io.File;
 import java.net.URL;
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
